refactor: extract sidebar into standalone component
